# api/openapi.yaml
openapi: 3.0.3
info:
  title: VimeSrv API
  description: REST API for VimeSrv - Home network video media server with DASH/HLS transcoding
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API token obtained from /api/v1/auth/login
    StreamAuth:
      type: apiKey
      in: query
      name: token
      description: Short-lived stream token obtained from /api/v1/auth/stream-token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_REQUEST"
            message:
              type: string
              example: "Invalid request body"
            details:
              type: object
...
    CandidatesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MetadataCandidate'
    MetadataCandidate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tmdb_id:
          type: integer
        media_type:
          type: string
          enum: [movie, tv]
        title:
          type: string
        release_date:
          type: string
        confidence:
          type: number
          description: Match confidence score (0-100)
        poster_url:
          type: string
    LinkRequest:
      type: object
      required:
        - candidate_id
      properties:
        candidate_id:
          type: integer
          format: int64
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          example: "The Matrix"
        year:
          type: integer
          example: 1999
        media_type:
          type: string
          enum: [movie, tv]
          description: Filter by media type (optional)
        max_results:
          type: integer
          default: 10
        language:
          type: string
          description: Language code for search results (e.g., "en", "de"). Falls back to server default if not provided.
          example: "en"
    SearchResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            media_id:
              type: string
            query:
              type: string
            results:
              type: array
              items:
                $ref: '#/components/schemas/SearchResult'
            count:
              type: integer
    SearchResult:
      type: object
      properties:
        tmdb_id:
          type: integer
        media_type:
          type: string
          enum: [movie, tv]
        title:
          type: string
        original_title:
          type: string
        release_date:
          type: string
        overview:
          type: string
        poster_url:
          type: string
    LinkSearchRequest:
      type: object
      required:
        - tmdb_id
        - media_type
      properties:
        tmdb_id:
          type: integer
        media_type:
          type: string
          enum: [movie, tv]
        season_number:
          type: integer
          description: For TV episodes
        episode_number:
          type: integer
          description: For TV episodes
...
    SuccessResponse:
      type: object
      properties:
        data:
          type: object

    # Library Browsing Schemas
...
paths:
...
  /api/v1/media/{id}/candidates:
    get:
      tags: [Metadata]
      summary: Get metadata candidates
      description: Returns TMDB metadata candidates for a media file
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: pending_only
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Metadata candidates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidatesResponse'
        '404':
          description: Media not found or no candidates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: TMDB integration disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/media/{id}/link:
    post:
      tags: [Metadata]
      summary: Link to metadata candidate
      description: Link a media file to a specific metadata candidate
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkRequest'
      responses:
        '200':
          description: Metadata linked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request or link failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: TMDB integration disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/media/{id}/search:
    post:
      tags: [Metadata]
      summary: Search TMDB
      description: Manually search TMDB for metadata
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: TMDB integration disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/media/{id}/link-search:
    post:
      tags: [Metadata]
      summary: Link from search result
      description: Link a media file directly from a TMDB search result
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkSearchRequest'
      responses:
        '200':
          description: Metadata linked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request or link failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: TMDB integration disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/media/{id}/skip:
    post:
      tags: [Metadata]
      summary: Skip metadata enrichment
      description: Mark a media file to skip metadata enrichment
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Enrichment skipped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Skip failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: TMDB integration disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/media/{id}/reset:
    post:
      tags: [Metadata]
      summary: Reset metadata enrichment
      description: Reset metadata enrichment status for a media file
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Enrichment reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Reset failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: TMDB integration disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...